+++ src/frontend/apps/e2e/__tests__/app-impress/404.spec.ts
@@ -17,13 +17,13 @@ test.describe('404', () => {
         'It seems that the page you are looking for does not exist or cannot be displayed correctly.',
       ),
     ).toBeVisible();
-    await expect(page.getByText('Back to home page')).toBeVisible();
+    await expect(page.getByText('Home')).toBeVisible();
   });
 
   test('checks go back to home page redirects to home page', async ({
     page,
   }) => {
-    await page.getByText('Back to home page').click();
+    await page.getByText('Home').click();
     await expect(page).toHaveURL('/');
   });
 });