+++ src/frontend/apps/impress/src/cunningham/cunningham-tokens.ts
@@ -505,6 +505,7 @@ export const tokens = {
               'color-active': 'var(--c--theme--colors--primary-100)',
             },
             'color-hover': 'var(--c--theme--colors--primary-text)',
+            color: 'var(--c--theme--colors--primary-800)',
           },
           secondary: {
             background: { 'color-hover': '#F6F6F6', 'color-active': '#EDEDED' },