+++ src/frontend/apps/impress/src/features/language/LanguagePicker.tsx
@@ -1,41 +1,16 @@
-import { Select } from '@openfun/cunningham-react';
 import { Settings } from 'luxon';
 import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import styled from 'styled-components';
+import { css } from 'styled-components';
 
-import { Box, Text } from '@/components/';
+import { DropdownMenu, Text } from '@/components/';
 import { LANGUAGES_ALLOWED } from '@/i18n/conf';
-
-const SelectStyled = styled(Select)<{ $isSmall?: boolean }>`
-  flex-shrink: 0;
-  width: auto;
-
-  .c__select__wrapper {
-    min-height: 2rem;
-    height: auto;
-    border-color: transparent;
-    padding: 0 0.15rem 0 0.45rem;
-    border-radius: 1px;
-
-    .labelled-box .labelled-box__children {
-      padding-right: 2rem;
-
-      .c_select__render .typo-text {
-        ${({ $isSmall }) => $isSmall && `display: none;`}
-      }
-    }
-
-    &:hover {
-      box-shadow: none !important;
-    }
-  }
-`;
 
 export const LanguagePicker = () => {
   const { t, i18n } = useTranslation();
   const { preload: languages } = i18n.options;
-  Settings.defaultLocale = i18n.language;
+  const language = i18n.language;
+  Settings.defaultLocale = language;
 
   const optionsPicker = useMemo(() => {
     return (languages || []).map((lang) => ({
@@ -62,23 +37,48 @@ export const LanguagePicker = () => {
           </Text>
         </Box>
       ),
-    }));
-  }, [languages]);
-
-  return (
-    <SelectStyled
-      label={t('Language')}
-      showLabelWhenSelected={false}
-      clearable={false}
-      hideLabel
-      defaultValue={i18n.language}
-      className="c_select__no_bg"
-      options={optionsPicker}
-      onChange={(e) => {
-        i18n.changeLanguage(e.target.value as string).catch((err) => {
+      label: LANGUAGES_ALLOWED[lang],
+      isSelected: language === lang,
+      callback: () => {
+        i18n.changeLanguage(lang).catch((err) => {
           console.error('Error changing language', err);
         });
-      }}
-    />
+      },
+    }));
+  }, [i18n, language, languages]);
+
+  return (
+    <DropdownMenu
+      options={optionsPicker}
+      showArrow
+      buttonCss={css`
+        &:hover {
+          background-color: var(
+            --c--components--button--primary-text--background--color-hover
+          );
+        }
+        border-radius: 4px;
+        padding: 0.5rem 0.6rem;
+        & > div {
+          gap: 0.2rem;
+          display: flex;
+        }
+        & .material-icons {
+          color: var(--c--components--button--primary-text--color) !important;
+        }
+      `}
+    >
+      <Text
+        $theme="primary"
+        aria-label={t('Language')}
+        $direction="row"
+        $gap="0.5rem"
+      >
+        <Text $isMaterialIcon $color="inherit" $size="xl">
+          translate
+        </Text>
+        {LANGUAGES_ALLOWED[language]}
+      </Text>
+    </DropdownMenu>
   );
 };