+++ src/frontend/apps/impress/src/layouts/PageLayout.tsx
@@ -6,17 +6,27 @@ import { HEADER_HEIGHT, Header } from '@/features/header';
 import { LeftPanel } from '@/features/left-panel';
 import { useResponsiveStore } from '@/stores';
 
-export function PageLayout({ children }: PropsWithChildren) {
+interface PageLayoutProps {
+  withFooter?: boolean;
+}
+
+export function PageLayout({
+  children,
+  withFooter = true,
+}: PropsWithChildren<PageLayoutProps>) {
   const { isDesktop } = useResponsiveStore();
 
   return (
-    <Box $minHeight="100vh" $margin={{ top: `${HEADER_HEIGHT}px` }}>
+    <Box
+      $minHeight={`calc(100vh - ${HEADER_HEIGHT}px)`}
+      $margin={{ top: `${HEADER_HEIGHT}px` }}
+    >
       <Header />
       <Box as="main" $width="100%" $css="flex-grow:1;">
         {!isDesktop && <LeftPanel />}
         {children}
       </Box>
-      <Footer />
+      {withFooter && <Footer />}
     </Box>
   );
 }