+++ src/frontend/apps/impress/src/pages/403.tsx
@@ -1,37 +1,42 @@
 import { Button } from '@openfun/cunningham-react';
+import Image from 'next/image';
 import { ReactElement } from 'react';
 import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
 
-import Icon404 from '@/assets/icons/icon-404.svg';
+import img403 from '@/assets/icons/icon-403.png';
 import { Box, StyledLink, Text } from '@/components';
 import { PageLayout } from '@/layouts';
 import { NextPageWithLayout } from '@/types/next';
 
 const StyledButton = styled(Button)`
   width: fit-content;
-  padding-left: 2rem;
-  padding-right: 2rem;
 `;
 
 const Page: NextPageWithLayout = () => {
   const { t } = useTranslation();
 
   return (
-    <Box $align="center" $margin="auto" $height="70vh" $gap="2rem">
-      <Icon404 aria-label="Image 404" role="img" />
+    <Box
+      $align="center"
+      $margin="auto"
+      $gap="1rem"
+      $padding={{ bottom: '2rem' }}
+    >
+      <Image
+        src={img403}
+        alt={t('Image 403')}
+        style={{
+          maxWidth: '100%',
+          height: 'auto',
+        }}
+      />
 
-      <Text $size="h2" $weight="700" $theme="greyscale" $variation="900">
-        {t('Ouch !')}
-      </Text>
+      <Box $align="center" $gap="0.8rem">
+        <Text as="p" $textAlign="center" $maxWidth="350px" $theme="primary">
+          {t('You do not have permission to view this document.')}
+        </Text>
 
-      <Text as="p" $textAlign="center" $maxWidth="400px" $size="m">
-        {t(
-          'It seems that the page you are looking for does not exist or cannot be displayed correctly.',
-        )}
-      </Text>
-
-      <Box $margin={{ top: 'large' }}>
         <StyledLink href="/">
           <StyledButton
             icon={